pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-inspector"
version = "0.1.0"
description = "Prompt-engineering secure code inspector for OWASP Top 10 mapping"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openai>=1.40.0",
  "PyYAML>=6.0.1",
  "pydantic>=2.8.0",
  "streamlit>=1.37.0",
  "semgrep>=1.152.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
